pyproject.toml
[tool.poetry]
name = "coreason_ai_gateway"
version = "0.2.0"
description = "A new Python project."
authors = ["Gowtham A Rao <gowtham.rao@coreason.ai>"]
license = "Prosperity-3.0"
readme = "README.md"
packages = [{include = "coreason_ai_gateway", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.12, <3.15"
loguru = "^0.7.2"
fastapi = "^0.128.0"
uvicorn = {extras = ["standard"], version = "^0.40.0"}
openai = "^2.15.0"
tenacity = "^9.1.2"
redis = "^7.1.0"
pydantic-settings = "^2.12.0"
coreason-vault = "^0.3.0"
coreason-identity = "^0.1.0"
anyio = "^4.12.1"
httpx = "^0.28.0"
aiofiles = "^25.1.0"

[tool.poetry.group.dev.dependencies]
types-aiofiles = "^25.1.0"
pytest-asyncio = "^0.23.0"
pytest = "^8.2.2"
ruff = "^0.4.8"
pre-commit = "^3.7.1"
pytest-cov = "^5.0.0"
mkdocs = "^1.6.0"
mkdocs-material = "^9.5.26"
respx = "^0.22.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[project]
name = "coreason_ai_gateway"
version = "0.1.0"
description = "A new Python project."
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Gowtham A Rao", email = "gowtham.rao@coreason.ai" },
]
license = { file = "LICENSE" }
classifiers = [
    "License :: Other/Proprietary License",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://github.com/CoReason-AI/coreason_ai_gateway"
Repository = "https://github.com/CoReason-AI/coreason_ai_gateway"
Documentation = "https://github.com/CoReason-AI/coreason_ai_gateway"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "B", "I"]
ignore = []

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "--cov=src --cov-report=term-missing --cov-report=html --cov-fail-under=100"
testpaths = ["tests"]

[tool.coverage.run]
omit = ["tests/*"]
